@@ -98,64 +98,64 @@
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-transform-runtime": "^7.17.0",
     "@babel/preset-env": "^7.16.11",
-    "@babel/register": "^7.17.0",
-    "@babel/runtime": "^7.17.2",
-    "@markedjs/html-differ": "^4.0.0",
-    "@playwright/test": "^1.19.1",
-    "babel-loader": "^8.2.3",
+    "@babel/register": "^7.17.7",
+    "@babel/runtime": "^7.17.9",
+    "@markedjs/html-differ": "^3.0.4",
+    "@playwright/test": "^1.21.0",
+    "babel-loader": "^8.2.4",
     "babel-plugin-component": "^1.1.1",
     "babel-plugin-istanbul": "^6.1.1",
-    "cfonts": "^2.10.0",
+    "cfonts": "^2.10.1",
     "chai": "^4.3.6",
     "chalk": "^4.1.2",
     "cheerio": "^1.0.0-rc.10",
     "copy-webpack-plugin": "^10.2.4",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.6.0",
+    "css-loader": "^6.7.1",
     "del": "^6.0.0",
     "devtron": "^1.4.0",
-    "dotenv": "^10.0.0",
-    "electron": "^15.4.0",
-    "electron-builder": "^22.14.13",
+    "dotenv": "^16.0.0",
+    "electron": "^18.0.4",
+    "electron-builder": "^23.0.6",
     "electron-devtools-installer": "^3.2.0",
     "electron-rebuild": "^3.2.7",
-    "electron-updater": "^4.6.5",
-    "eslint": "^8.9.0",
+    "electron-updater": "^5.0.2",
+    "eslint": "^8.13.0",
     "eslint-config-standard": "^16.0.3",
     "eslint-friendly-formatter": "^4.0.1",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-html": "^6.2.0",
-    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.0.0",
     "eslint-plugin-standard": "^4.1.0",
-    "eslint-plugin-vue": "^8.4.1",
+    "eslint-plugin-vue": "^8.6.0",
     "eslint-webpack-plugin": "^3.1.1",
     "esm": "^3.2.25",
     "file-loader": "^6.2.0",
     "git-revision-webpack-plugin": "^5.0.0",
     "html-webpack-plugin": "^5.5.0",
     "imports-loader": "^0.8.0",
-    "karma": "^6.3.16",
+    "karma": "^6.3.18",
     "karma-chai": "^0.1.0",
     "karma-coverage": "^2.2.0",
     "karma-electron": "^7.1.0",
     "karma-mocha": "^2.0.1",
     "karma-sourcemap-loader": "^0.3.8",
-    "karma-spec-reporter": "0.0.33",
+    "karma-spec-reporter": "0.0.34",
     "karma-webpack": "^5.0.0",
     "license-checker": "^25.0.1",
     "listr": "^0.14.3",
     "marked": "^1.2.9",
-    "mini-css-extract-plugin": "^2.5.3",
-    "mocha": "^8.4.0",
-    "node-fetch": "^2.6.6",
+    "mini-css-extract-plugin": "^2.6.0",
+    "mocha": "^9.2.2",
+    "node-fetch": "^2.6.7",
     "node-loader": "^2.0.0",
     "path-browserify": "^1.0.1",
-    "playwright": "^1.19.1",
-    "postcss": "^8.4.6",
+    "playwright": "^1.21.0",
+    "postcss": "^8.4.12",
     "postcss-loader": "^6.2.1",
-    "postcss-preset-env": "^7.4.1",
+    "postcss-preset-env": "^7.4.3",
     "raw-loader": "^4.0.2",
     "require-dir": "^1.2.0",
     "stacktrace-parser": "^0.1.10",
@@ -169,16 +169,16 @@
     "vue-loader": "^15.9.8",
     "vue-style-loader": "^4.1.3",
     "vue-template-compiler": "^2.6.14",
-    "webpack": "^5.69.1",
+    "webpack": "^5.72.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4",
+    "webpack-dev-server": "^4.8.1",
     "webpack-hot-middleware": "^2.25.1",
     "webpack-merge": "^5.8.0"
   },
   "resolutions": {
-    "cld": "^2.7.1",
-    "node-abi": "^3.5.0"
+    "node-abi": "^3.8.0",
+    "node-addon-api": "^4.3.0"
   },
   "repository": {
     "type": "git",
