@@ -2,7 +2,7 @@ module.exports = {
   root: true,
   parserOptions: {
     parser: '@babel/eslint-parser',
-    ecmaVersion: 8,
+    ecmaVersion: 11,
     ecmaFeatures: {
       impliedStrict: true
     },
@@ -25,29 +25,23 @@ module.exports = {
   },
   plugins: ['html', 'vue'],
   rules: {
+    // Two spaces but disallow semicolons
+    indent: ['error', 2, { 'SwitchCase': 1, 'ignoreComments': true }],
+    semi: [2, 'never'],
+    'no-return-await': 'error',
+    'no-return-assign': 'error',
+    'no-new': 'error',
     // allow paren-less arrow functions
     'arrow-parens': 'off',
-    // allow async-await
-    'generator-star-spacing': 'off',
     // allow console
     'no-console': 'off',
     // allow debugger during development
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    'no-return-assign': 'warn',
-    'no-new': 'warn',
-    // disallow semicolons
-    semi: [2, 'never'],
     'require-atomic-updates': 'off',
     // TODO: fix these errors someday
     'prefer-const': 'off',
-    'no-new': 'off',
     'no-mixed-operators': 'off',
-    'no-prototype-builtins': 'off',
-    'no-return-await': 'off',
-    'accessor-pairs': 'off',
-    // Workaround #2422.
-    'template-curly-spacing': 'off',
-    indent: 'off'
+    'no-prototype-builtins': 'off'
   },
   settings: {
     'import/resolver': {
