@@ -3,6 +3,8 @@ import Preference from '../preferences'
 import DataCenter from '../dataCenter'
 import Keybindings from '../keyboard/shortcutHandler'
 import AppMenu from '../menu'
+import { loadMenuCommands } from '../menu/actions'
+import { CommandManager, loadDefaultCommands } from '../commands'
 
 class Accessor {
   /**
@@ -13,11 +15,26 @@ class Accessor {
 
     this.env = appEnvironment
     this.paths = appEnvironment.paths // export paths to make it better accessible
+
     this.preferences = new Preference(this.paths)
     this.dataCenter = new DataCenter(this.paths)
-    this.keybindings = new Keybindings(userDataPath)
+
+    this.commandManager = CommandManager
+    this._loadCommands()
+
+    this.keybindings = new Keybindings(this.commandManager, appEnvironment)
     this.menu = new AppMenu(this.preferences, this.keybindings, userDataPath)
     this.windowManager = new WindowManager(this.menu, this.preferences)
+  }
+
+  _loadCommands () {
+    const { commandManager } = this
+    loadDefaultCommands(commandManager)
+    loadMenuCommands(commandManager)
+
+    if (this.env.isDevMode) {
+      commandManager.__verifyDefaultCommands()
+    }
   }
 }
 
