@@ -14,8 +14,10 @@ export class AppEnvironment {
     this._id = envId++
     this._appPaths = new AppPaths(options.userDataPath)
     this._debug = !!options.debug
+    this._isDevMode = !!options.isDevMode
     this._verbose = !!options.verbose
     this._safeMode = !!options.safeMode
+    this._disableSpellcheck = !!options.disableSpellcheck
   }
 
   /**
@@ -44,6 +46,13 @@ export class AppEnvironment {
   /**
    * @returns {boolean}
    */
+  get isDevMode () {
+    return this._isDevMode
+  }
+
+  /**
+   * @returns {boolean}
+   */
   get verbose () {
     return this._verbose
   }
@@ -53,6 +62,13 @@ export class AppEnvironment {
    */
   get safeMode () {
     return this._safeMode
+  }
+
+  /**
+   * @returns {boolean}
+   */
+  get disableSpellcheck () {
+    return this._disableSpellcheck
   }
 }
 
@@ -65,16 +81,20 @@ export class AppEnvironment {
 const setupEnvironment = args => {
   patchEnvPath()
 
+  const isDevMode = process.env.NODE_ENV !== 'production'
   const debug = args['--debug'] || !!process.env.MARKTEXT_DEBUG || process.env.NODE_ENV !== 'production'
   const verbose = args['--verbose'] || 0
   const safeMode = args['--safe']
   const userDataPath = args['--user-data-dir'] // or null (= default user data path)
+  const disableSpellcheck = args['--disable-spellcheck']
 
   const appEnvironment = new AppEnvironment({
     debug,
+    isDevMode,
     verbose,
     safeMode,
-    userDataPath
+    userDataPath,
+    disableSpellcheck
   })
 
   ensureAppDirectoriesSync(appEnvironment.paths)
