@@ -1,4 +1,6 @@
-const MENU_ID_FORMAT_MAP = {
+import { COMMANDS } from '../../commands'
+
+const MENU_ID_FORMAT_MAP = Object.freeze({
   strongMenuItem: 'strong',
   emphasisMenuItem: 'em',
   inlineCodeMenuItem: 'inline_code',
@@ -6,12 +8,77 @@ const MENU_ID_FORMAT_MAP = {
   hyperlinkMenuItem: 'link',
   imageMenuItem: 'image',
   inlineMathMenuItem: 'inline_math'
-}
+})
 
-export const format = (win, type) => {
+const format = (win, type) => {
   if (win && win.webContents) {
     win.webContents.send('mt::editor-format-action', { type })
   }
+}
+
+export const clearFormat = win => {
+  format(win, 'clear')
+}
+
+export const emphasis = win => {
+  format(win, 'em')
+}
+
+export const highlight = win => {
+  format(win, 'mark')
+}
+
+export const hyperlink = win => {
+  format(win, 'link')
+}
+
+export const image = win => {
+  format(win, 'image')
+}
+
+export const inlineCode = win => {
+  format(win, 'inline_code')
+}
+
+export const inlineMath = win => {
+  format(win, 'inline_math')
+}
+
+export const strikethrough = win => {
+  format(win, 'del')
+}
+
+export const strong = win => {
+  format(win, 'strong')
+}
+
+export const subscript = win => {
+  format(win, 'sub')
+}
+
+export const superscript = win => {
+  format(win, 'sup')
+}
+
+export const underline = win => {
+  format(win, 'u')
+}
+
+// --- Commands -------------------------------------------------------------
+
+export const loadFormatCommands = commandManager => {
+  commandManager.add(COMMANDS.FORMAT_CLEAR_FORMAT, clearFormat)
+  commandManager.add(COMMANDS.FORMAT_EMPHASIS, emphasis)
+  commandManager.add(COMMANDS.FORMAT_HIGHLIGHT, highlight)
+  commandManager.add(COMMANDS.FORMAT_HYPERLINK, hyperlink)
+  commandManager.add(COMMANDS.FORMAT_IMAGE, image)
+  commandManager.add(COMMANDS.FORMAT_INLINE_CODE, inlineCode)
+  commandManager.add(COMMANDS.FORMAT_INLINE_MATH, inlineMath)
+  commandManager.add(COMMANDS.FORMAT_STRIKE, strikethrough)
+  commandManager.add(COMMANDS.FORMAT_STRONG, strong)
+  commandManager.add(COMMANDS.FORMAT_SUBSCRIPT, subscript)
+  commandManager.add(COMMANDS.FORMAT_SUPERSCRIPT, superscript)
+  commandManager.add(COMMANDS.FORMAT_UNDERLINE, underline)
 }
 
 // --- IPC events -------------------------------------------------------------
