@@ -1,5 +1,7 @@
 import { autoUpdater } from 'electron-updater'
-import { ipcMain, BrowserWindow } from 'electron'
+import { ipcMain, BrowserWindow, Menu } from 'electron'
+import { COMMANDS } from '../../commands'
+import { isOsx } from '../../config'
 
 let runningUpdate = false
 let win = null
@@ -36,18 +38,6 @@ autoUpdater.on('update-downloaded', () => {
   setImmediate(() => autoUpdater.quitAndInstall())
 })
 
-export const userSetting = () => {
-  ipcMain.emit('app-create-settings-window')
-}
-
-export const checkUpdates = browserWindow => {
-  if (!runningUpdate) {
-    runningUpdate = true
-    win = browserWindow
-    autoUpdater.checkForUpdates()
-  }
-}
-
 ipcMain.on('mt::NEED_UPDATE', (e, { needUpdate }) => {
   if (needUpdate) {
     autoUpdater.downloadUpdate()
@@ -60,3 +50,42 @@ ipcMain.on('mt::check-for-update', e => {
   const win = BrowserWindow.fromWebContents(e.sender)
   checkUpdates(win)
 })
+
+// --------------------------------------------------------
+
+export const userSetting = () => {
+  ipcMain.emit('app-create-settings-window')
+}
+
+export const checkUpdates = browserWindow => {
+  if (!runningUpdate) {
+    runningUpdate = true
+    win = browserWindow
+    autoUpdater.checkForUpdates()
+  }
+}
+
+export const osxHide = () => {
+  if (isOsx) {
+    Menu.sendActionToFirstResponder('hide:')
+  }
+}
+
+export const osxHideAll = () => {
+  if (isOsx) {
+    Menu.sendActionToFirstResponder('hideOtherApplications:')
+  }
+}
+
+export const osxShowAll = () => {
+  if (isOsx) {
+    Menu.sendActionToFirstResponder('unhideAllApplications:')
+  }
+}
+
+// --- Commands -------------------------------------------------------------
+
+export const loadMarktextCommands = commandManager => {
+  commandManager.add(COMMANDS.MT_HIDE, osxHide)
+  commandManager.add(COMMANDS.MT_HIDE_OTHERS, osxHideAll)
+}
