@@ -1,38 +1,93 @@
-import { getMenuItemById } from '../../menu'
+import { ipcMain } from 'electron'
+import { COMMANDS } from '../../commands'
 
 const typewriterModeMenuItemId = 'typewriterModeMenuItem'
 const focusModeMenuItemId = 'focusModeMenuItem'
 
-export const showCommandPalette = win => {
-  win.webContents.send('mt::show-command-palette')
-}
-
-export const typeMode = (win, type, item) => {
-  if (!win) {
-    return
-  }
-  const { checked } = item
-  win.webContents.send('mt::editor-change-view', { type, checked })
-
-  if (type === 'sourceCode') {
-    const typewriterModeMenuItem = getMenuItemById(typewriterModeMenuItemId)
-    const focusModeMenuItem = getMenuItemById(focusModeMenuItemId)
-    typewriterModeMenuItem.enabled = !checked
-    focusModeMenuItem.enabled = !checked
-  }
-}
-
-export const layout = (item, win, type, value) => {
+const toggleTypeMode = (win, type) => {
   if (win && win.webContents) {
-    win.webContents.send('mt::set-view-layout', { [type]: value || item.checked })
+    win.webContents.send('mt::toggle-view-mode-entry', type)
   }
+}
+
+const setLayout = (win, type, value) => {
+  if (win && win.webContents) {
+    win.webContents.send('mt::set-view-layout', { [type]: value })
+  }
+}
+
+const toggleLayout = (win, type) => {
+  if (win && win.webContents) {
+    win.webContents.send('mt::toggle-view-layout-entry', type)
+  }
+}
+
+export const debugToggleDevTools = win => {
+  if (win && global.MARKTEXT_DEBUG) {
+    win.webContents.toggleDevTools()
+  }
+}
+
+export const debugReloadWindow = win => {
+  if (win && global.MARKTEXT_DEBUG) {
+    ipcMain.emit('window-reload-by-id', win.id)
+  }
+}
+
+export const showCommandPalette = win => {
+  if (win && win.webContents) {
+    win.webContents.send('mt::show-command-palette')
+  }
+}
+
+export const toggleFocusMode = win => {
+  toggleTypeMode(win, 'focus')
+}
+
+export const toggleSourceCodeMode = win => {
+  toggleTypeMode(win, 'sourceCode')
+}
+
+export const toggleSidebar = win => {
+  toggleLayout(win, 'showSideBar')
+}
+
+export const toggleTabBar = win => {
+  toggleLayout(win, 'showTabBar')
 }
 
 export const showTabBar = win => {
-  const tabBarMenuItem = getMenuItemById('tabBarMenuItem')
-  if (tabBarMenuItem && !tabBarMenuItem.checked && tabBarMenuItem.click) {
-    tabBarMenuItem.click(tabBarMenuItem, win)
+  setLayout(win, 'showTabBar', true)
+}
+
+export const showTableOfContents = win => {
+  setLayout(win, 'rightColumn', 'toc')
+}
+
+export const toggleTypewriterMode = win => {
+  toggleTypeMode(win, 'typewriter')
+}
+
+export const reloadImageCache = win => {
+  if (win && win.webContents) {
+    win.webContents.send('mt::invalidate-image-cache')
   }
+}
+
+// --- Commands -------------------------------------------------------------
+
+export const loadViewCommands = commandManager => {
+  commandManager.add(COMMANDS.VIEW_COMMAND_PALETTE, showCommandPalette)
+  commandManager.add(COMMANDS.VIEW_FOCUS_MODE, toggleFocusMode)
+  commandManager.add(COMMANDS.VIEW_FORCE_RELOAD_IMAGES, reloadImageCache)
+  commandManager.add(COMMANDS.VIEW_SOURCE_CODE_MODE, toggleSourceCodeMode)
+  commandManager.add(COMMANDS.VIEW_TOGGLE_SIDEBAR, toggleSidebar)
+  commandManager.add(COMMANDS.VIEW_TOGGLE_TABBAR, toggleTabBar)
+  commandManager.add(COMMANDS.VIEW_TOGGLE_TOC, showTableOfContents)
+  commandManager.add(COMMANDS.VIEW_TYPEWRITER_MODE, toggleTypewriterMode)
+
+  commandManager.add(COMMANDS.VIEW_DEV_RELOAD, debugReloadWindow)
+  commandManager.add(COMMANDS.VIEW_TOGGLE_DEV_TOOLS, debugToggleDevTools)
 }
 
 // --- IPC events -------------------------------------------------------------
@@ -46,6 +101,10 @@ export const showTabBar = win => {
  * @param {*} changes Array of changed view settings (e.g. [ {showSideBar: true} ]).
  */
 export const viewLayoutChanged = (applicationMenu, changes) => {
+  const disableMenuByName = (id, value) => {
+    const menuItem = applicationMenu.getMenuItemById(id)
+    menuItem.enabled = value
+  }
   const changeMenuByName = (id, value) => {
     const menuItem = applicationMenu.getMenuItemById(id)
     menuItem.checked = value
@@ -62,6 +121,8 @@ export const viewLayoutChanged = (applicationMenu, changes) => {
         break
       case 'sourceCode':
         changeMenuByName('sourceCodeModeMenuItem', value)
+        disableMenuByName(focusModeMenuItemId, !value)
+        disableMenuByName(typewriterModeMenuItemId, !value)
         break
       case 'typewriter':
         changeMenuByName(typewriterModeMenuItemId, value)
