@@ -9,24 +9,24 @@ export default function (keybindings) {
       label: 'Bold',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.strong'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'strong')
+      click (menuItem, focusedWindow) {
+        actions.strong(focusedWindow)
       }
     }, {
       id: 'emphasisMenuItem',
       label: 'Italic',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.emphasis'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'em')
+      click (menuItem, focusedWindow) {
+        actions.emphasis(focusedWindow)
       }
     }, {
       id: 'underlineMenuItem',
       label: 'Underline',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.underline'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'u')
+      click (menuItem, focusedWindow) {
+        actions.underline(focusedWindow)
       }
     }, {
       type: 'separator'
@@ -35,24 +35,24 @@ export default function (keybindings) {
       label: 'Superscript',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.superscript'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'sup')
+      click (menuItem, focusedWindow) {
+        actions.superscript(focusedWindow)
       }
     }, {
       id: 'subscriptMenuItem',
       label: 'Subscript',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.subscript'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'sub')
+      click (menuItem, focusedWindow) {
+        actions.subscript(focusedWindow)
       }
     }, {
       id: 'highlightMenuItem',
       label: 'Highlight',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.highlight'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'mark')
+      click (menuItem, focusedWindow) {
+        actions.highlight(focusedWindow)
       }
     }, {
       type: 'separator'
@@ -61,16 +61,16 @@ export default function (keybindings) {
       label: 'Inline Code',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.inline-code'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'inline_code')
+      click (menuItem, focusedWindow) {
+        actions.inlineCode(focusedWindow)
       }
     }, {
       id: 'inlineMathMenuItem',
       label: 'Inline Math',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.inline-math'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'inline_math')
+      click (menuItem, focusedWindow) {
+        actions.inlineMath(focusedWindow)
       }
     }, {
       type: 'separator'
@@ -79,32 +79,32 @@ export default function (keybindings) {
       label: 'Strikethrough',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.strike'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'del')
+      click (menuItem, focusedWindow) {
+        actions.strikethrough(focusedWindow)
       }
     }, {
       id: 'hyperlinkMenuItem',
       label: 'Hyperlink',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.hyperlink'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'link')
+      click (menuItem, focusedWindow) {
+        actions.hyperlink(focusedWindow)
       }
     }, {
       id: 'imageMenuItem',
       label: 'Image',
       type: 'checkbox',
       accelerator: keybindings.getAccelerator('format.image'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'image')
+      click (menuItem, focusedWindow) {
+        actions.image(focusedWindow)
       }
     }, {
       type: 'separator'
     }, {
       label: 'Clear Formatting',
       accelerator: keybindings.getAccelerator('format.clear-format'),
-      click (menuItem, browserWindow) {
-        actions.format(browserWindow, 'clear')
+      click (menuItem, focusedWindow) {
+        actions.clearFormat(focusedWindow)
       }
     }]
   }
