@@ -1,4 +1,3 @@
-import { ipcMain } from 'electron'
 import * as actions from '../actions/view'
 
 export default function (keybindings) {
@@ -7,8 +6,8 @@ export default function (keybindings) {
     submenu: [{
       label: 'Command Palette...',
       accelerator: keybindings.getAccelerator('view.command-palette'),
-      click (menuItem, browserWindow) {
-        actions.showCommandPalette(browserWindow)
+      click (menuItem, focusedWindow) {
+        actions.showCommandPalette(focusedWindow)
       }
     }, {
       type: 'separator'
@@ -18,12 +17,8 @@ export default function (keybindings) {
       accelerator: keybindings.getAccelerator('view.source-code-mode'),
       type: 'checkbox',
       checked: false,
-      click (item, browserWindow, event) {
-        // if we call this function, the checked state is not set
-        if (!event) {
-          item.checked = !item.checked
-        }
-        actions.typeMode(browserWindow, 'sourceCode', item)
+      click (item, focusedWindow) {
+        actions.toggleSourceCodeMode(focusedWindow)
       }
     }, {
       id: 'typewriterModeMenuItem',
@@ -31,12 +26,8 @@ export default function (keybindings) {
       accelerator: keybindings.getAccelerator('view.typewriter-mode'),
       type: 'checkbox',
       checked: false,
-      click (item, browserWindow, event) {
-        // if we call this function, the checked state is not set
-        if (!event) {
-          item.checked = !item.checked
-        }
-        actions.typeMode(browserWindow, 'typewriter', item)
+      click (item, focusedWindow) {
+        actions.toggleTypewriterMode(focusedWindow)
       }
     }, {
       id: 'focusModeMenuItem',
@@ -44,12 +35,8 @@ export default function (keybindings) {
       accelerator: keybindings.getAccelerator('view.focus-mode'),
       type: 'checkbox',
       checked: false,
-      click (item, browserWindow, event) {
-        // if we call this function, the checked state is not set
-        if (!event) {
-          item.checked = !item.checked
-        }
-        actions.typeMode(browserWindow, 'focus', item)
+      click (item, focusedWindow) {
+        actions.toggleFocusMode(focusedWindow)
       }
     }, {
       type: 'separator'
@@ -59,13 +46,8 @@ export default function (keybindings) {
       accelerator: keybindings.getAccelerator('view.toggle-sidebar'),
       type: 'checkbox',
       checked: false,
-      click (item, browserWindow, event) {
-        // if we call this function, the checked state is not set
-        if (!event) {
-          item.checked = !item.checked
-        }
-
-        actions.layout(item, browserWindow, 'showSideBar')
+      click (item, focusedWindow) {
+        actions.toggleSidebar(focusedWindow)
       }
     }, {
       label: 'Show Tab Bar',
@@ -73,51 +55,41 @@ export default function (keybindings) {
       accelerator: keybindings.getAccelerator('view.toggle-tabbar'),
       type: 'checkbox',
       checked: false,
-      click (item, browserWindow, event) {
-        // if we call this function, the checked state is not set
-        if (!event) {
-          item.checked = !item.checked
-        }
-
-        actions.layout(item, browserWindow, 'showTabBar')
+      click (item, focusedWindow) {
+        actions.toggleTabBar(focusedWindow)
       }
     }, {
       label: 'Toggle Table of Contents',
       id: 'tocMenuItem',
       accelerator: keybindings.getAccelerator('view.toggle-toc'),
-      click (_, browserWindow) {
-        actions.layout(null, browserWindow, 'rightColumn', 'toc')
+      click (_, focusedWindow) {
+        actions.showTableOfContents(focusedWindow)
       }
     }, {
       label: 'Reload Images',
       accelerator: keybindings.getAccelerator('view.reload-images'),
       click (item, focusedWindow) {
-        if (focusedWindow) {
-          focusedWindow.webContents.send('mt::invalidate-image-cache', {})
-        }
+        actions.reloadImageCache(focusedWindow)
       }
-    }, {
-      type: 'separator'
     }]
   }
 
   if (global.MARKTEXT_DEBUG) {
     viewMenu.submenu.push({
+      type: 'separator'
+    })
+    viewMenu.submenu.push({
       label: 'Show Developer Tools',
       accelerator: keybindings.getAccelerator('view.toggle-dev-tools'),
-      click (item, focusedWindow) {
-        if (focusedWindow) {
-          focusedWindow.webContents.toggleDevTools()
-        }
+      click (item, win) {
+        actions.debugToggleDevTools(win)
       }
     })
     viewMenu.submenu.push({
       label: 'Reload window',
       accelerator: keybindings.getAccelerator('view.dev-reload'),
       click (item, focusedWindow) {
-        if (focusedWindow) {
-          ipcMain.emit('window-reload-by-id', focusedWindow.id)
-        }
+        actions.debugReloadWindow(focusedWindow)
       }
     })
   }
