@@ -13,8 +13,11 @@ class ClickEvent {
   contextClickBingding () {
     const { container, eventCenter, contentState } = this.muya
     const handler = event => {
-      event.preventDefault()
-      event.stopPropagation()
+      // Allow native context menu in MarkText.
+      if (!global || !global.marktext) { // __MARKTEXT_PATCH__
+        event.preventDefault()
+        event.stopPropagation()
+      }
 
       // Hide all float box and image transformer
       const { keyboard } = this.muya
