@@ -11,21 +11,21 @@ export default function htmlRuby (h, cursor, block, token, outerClass) {
   const previewSelector = `span.${CLASS_OR_ID.AG_RUBY_RENDER}`
 
   return children
-  ? [
-    h(`span.${className}.${CLASS_OR_ID.AG_RUBY}`, [
-      h(`span.${CLASS_OR_ID.AG_INLINE_RULE}.${CLASS_OR_ID.AG_RUBY_TEXT}`, content),
-      h(previewSelector, {
-        attrs: {
-          contenteditable: 'false',
-          spellcheck: 'false'
-        }
-      }, vNode)
-    ])
+    ? [
+      h(`span.${className}.${CLASS_OR_ID.AG_RUBY}`, [
+        h(`span.${CLASS_OR_ID.AG_INLINE_RULE}.${CLASS_OR_ID.AG_RUBY_TEXT}`, content),
+        h(previewSelector, {
+          attrs: {
+            contenteditable: 'false',
+            spellcheck: 'false'
+          }
+        }, vNode)
+      ])
   // if children is empty string, no need to render ruby charactors...
-  ]
-  : [
-    h(`span.${className}.${CLASS_OR_ID.AG_RUBY}`, [
-      h(`span.${CLASS_OR_ID.AG_INLINE_RULE}.${CLASS_OR_ID.AG_RUBY_TEXT}`, content)
-    ])
-  ]
+    ]
+    : [
+      h(`span.${className}.${CLASS_OR_ID.AG_RUBY}`, [
+        h(`span.${CLASS_OR_ID.AG_INLINE_RULE}.${CLASS_OR_ID.AG_RUBY_TEXT}`, content)
+      ])
+    ]
 }
