@@ -26,7 +26,6 @@ const parseUrlArgs = () => {
   const userDataPath = params.get('udp')
   const windowId = Number(params.get('wid'))
   const type = params.get('type')
-  const spellcheckerIsHunspell = params.get('slp') === '1'
 
   if (Number.isNaN(windowId)) {
     throw new Error('Error while parsing URL arguments: windowId!')
@@ -43,8 +42,7 @@ const parseUrlArgs = () => {
       hideScrollbar,
       theme,
       titleBarStyle
-    },
-    spellcheckerIsHunspell
+    }
   }
 }
 
@@ -73,8 +71,7 @@ const bootstrapRenderer = () => {
     initialState,
     userDataPath,
     windowId,
-    type,
-    spellcheckerIsHunspell
+    type
   } = parseUrlArgs()
   const paths = new RendererPaths(userDataPath)
   const marktext = {
@@ -88,14 +85,6 @@ const bootstrapRenderer = () => {
     paths
   }
   global.marktext = marktext
-
-  // Set option to always use Hunspell instead OS spell checker.
-  if (spellcheckerIsHunspell && type !== 'settings') {
-    // HACK: This code doesn't do anything because `node-spellchecker` is loaded by
-    // `internal/modules/cjs/loader.js` before we can set the envoriment variable here.
-    // The code is additionally added to `index.ejs` to workaound the problem.
-    process.env['SPELLCHECKER_PREFER_HUNSPELL'] = 1 // eslint-disable-line dot-notation
-  }
 
   configureLogger()
 }
