@@ -9,6 +9,7 @@ import getCommandDescriptionById from './descriptions'
 export { default as FileEncodingCommand } from './fileEncoding'
 export { default as LineEndingCommand } from './lineEnding'
 export { default as QuickOpenCommand } from './quickOpen'
+export { default as SpellcheckerLanguageCommand } from './spellcheckerLanguage'
 export { default as TrailingNewlineCommand } from './trailingNewline'
 
 export class RootCommand {
@@ -43,7 +44,7 @@ const commands = [
       ipcRenderer.emit('mt::new-untitled-tab', null)
     }
   }, {
-    id: 'file.new-file',
+    id: 'file.new-window',
     execute: async () => {
       ipcRenderer.send('mt::cmd-new-editor-window')
     }
@@ -581,12 +582,12 @@ const commands = [
   }, {
     id: 'view.toggle-sidebar',
     execute: async () => {
-      bus.$emit('view:toggle-view-layout-entry', 'showSideBar')
+      bus.$emit('view:toggle-layout-entry', 'showSideBar')
     }
   }, {
     id: 'view.toggle-tabbar',
     execute: async () => {
-      bus.$emit('view:toggle-view-layout-entry', 'showTabBar')
+      bus.$emit('view:toggle-layout-entry', 'showTabBar')
     }
   },
 
@@ -646,6 +647,9 @@ const commands = [
     }
   }
 ]
+
+// --------------------------------------------------------------------------
+// etc
 
 if (isUpdatable()) {
   commands.push({
