@@ -4,19 +4,12 @@
     <compound>
       <template #head>
         <bool
-          description="Enable spell checker"
+          description="Enable spell checking"
           :bool="spellcheckerEnabled"
           :onChange="handleSpellcheckerEnabled"
         ></bool>
       </template>
       <template #children>
-        <bool
-          description="Use Hunspell instead of system spell checker on macOS and Windows 10"
-          notes="Requires restart."
-          :bool="spellcheckerIsHunspell"
-          :disable="!isOsSpellcheckerSupported || !spellcheckerEnabled"
-          :onChange="value => onSelectChange('spellcheckerIsHunspell', value)"
-        ></bool>
         <bool
           description="Hide marks for spelling errors"
           :bool="spellcheckerNoUnderline"
@@ -24,93 +17,61 @@
           :onChange="value => onSelectChange('spellcheckerNoUnderline', value)"
         ></bool>
         <bool
-          v-show="isOsx && !spellcheckerIsHunspell"
-          description="Automatically detect document language (requires showing marks for spelling errors)"
-          :bool="spellcheckerAutoDetectLanguage"
-          :disable="!spellcheckerEnabled"
-          :onChange="value => onSelectChange('spellcheckerAutoDetectLanguage', value)"
+          v-show="isOsx"
+          description="Automatically detect document language"
+          :bool="true"
+          :disable="true"
         ></bool>
+        <cur-select
+          v-show="!isOsx"
+          description="Default language for spell checking"
+          :value="spellcheckerLanguage"
+          :options="availableDictionaries"
+          :disable="!spellcheckerEnabled"
+          :onChange="handleSpellcheckerLanguage"
+        ></cur-select>
       </template>
     </compound>
 
-    <separator></separator>
-
-    <cur-select
-      description="Default language for spell checker"
-      :value="spellcheckerLanguage"
-      :options="availableDictionaries"
-      :disable="!spellcheckerEnabled"
-      :onChange="value => onSelectChange('spellcheckerLanguage', value)"
-    ></cur-select>
-    <div
-      v-if="isOsx && !isHunspellSelected && spellcheckerEnabled"
-      class="description"
-    >
-      Additional languages may be added through "Language & Region" in your system preferences pane.
-    </div>
-    <div
-      v-if="isWindows && !isHunspellSelected && spellcheckerEnabled"
-      class="description"
-    >
-      Additional languages may be added through "Language" in your "Time & language" settings.
+    <div v-if="isOsx && spellcheckerEnabled" class="description">
+      The used language will be detected automatically while typing. Additional languages may be added through "Language & Region" in your system preferences pane.
     </div>
 
-    <div v-if="isHunspellSelected && spellcheckerEnabled">
-      <h6 class="title">Hunspell settings:</h6>
-      <div class="description">Installed Hunspell dictionaries</div>
+    <div v-if="!isOsx && spellcheckerEnabled">
+      <h6 class="title">Custom dictionary:</h6>
+      <div class="description">Edit words in custom dictionary.</div>
       <el-table
-        :data="availableDictionaries"
-        style="width: 100%">
-        <el-table-column
-          prop="value"
-          label="Name"
-          width="100">
+        :data="wordsInCustomDictionary"
+        empty-text="No words available"
+        style="width: 100%"
+      >
+        <el-table-column prop="word" label="Word">
         </el-table-column>
-        <el-table-column
-          prop="label"
-          label="Language"
-        >
-        </el-table-column>
-        <el-table-column
-          fixed="right"
-          label="Operations"
-          width="170">
+
+        <el-table-column fixed="right" label="Options" width="90">
           <template slot-scope="scope">
-            <el-button @click="handleUpdateClick(scope.$index, scope.row)" type="text" size="small">Update</el-button>
-            <el-button @click="handleDeleteClick(scope.$index, scope.row)" type="text" size="small">Delete</el-button>
+            <el-button @click="handleDeleteClick(scope.row)" type="text" size="small" title="Delete">
+              <i class="el-icon-delete"></i>
+            </el-button>
           </template>
         </el-table-column>
       </el-table>
-
-      <div class="description">Download additional Hunspell dictionaries</div>
-      <div class="dictionary-group">
-        <el-select
-          v-model="selectedDictionaryToAdd"
-        >
-          <el-option
-            v-for="item in dictionariesLanguagesOptions"
-            :key="item.value"
-            :label="item.label"
-            :value="item.value">
-          </el-option>
-        </el-select>
-        <el-button icon="el-icon-document-add" @click="addNewDict"></el-button>
-      </div>
-      <div v-if="errorMessage" class="description">{{ errorMessage }}</div>
     </div>
   </div>
 </template>
 
 <script>
+import { ipcRenderer } from 'electron'
+import log from 'electron-log'
 import { mapState } from 'vuex'
 import Compound from '../common/compound'
 import CurSelect from '../common/select'
 import Bool from '../common/bool'
 import Separator from '../common/separator'
-import { isOsx, isLinux, isWindows, cloneObj } from '@/util'
-import { isOsSpellcheckerSupported, getAvailableHunspellDictionaries, SpellChecker } from '@/spellchecker'
-import { getLanguageName, HUNSPELL_DICTIONARY_LANGUAGE_MAP } from '@/spellchecker/languageMap'
-import { downloadHunspellDictionary, deleteHunspellDictionary } from '@/spellchecker/dictionaryDownloader'
+import { isOsx } from '@/util'
+import { SpellChecker } from '@/spellchecker'
+import { getLanguageName } from '@/spellchecker/languageMap'
+import notice from '@/services/notification'
 
 export default {
   components: {
@@ -121,175 +82,85 @@ export default {
   },
   data () {
     this.isOsx = isOsx
-    this.isLinux = isLinux
-    this.isWindows = isWindows
-    this.isOsSpellcheckerSupported = isOsSpellcheckerSupported()
-    this.dictionariesLanguagesOptions = cloneObj(HUNSPELL_DICTIONARY_LANGUAGE_MAP)
-    this.hunspellDictionaryDownloadCache = {}
     return {
       availableDictionaries: [],
-      selectedDictionaryToAdd: 'en-US',
+      wordsInCustomDictionary: [],
       errorMessage: ''
     }
   },
   computed: {
     ...mapState({
       spellcheckerEnabled: state => state.preferences.spellcheckerEnabled,
-      spellcheckerIsHunspell: state => state.preferences.spellcheckerIsHunspell,
       spellcheckerNoUnderline: state => state.preferences.spellcheckerNoUnderline,
-      spellcheckerAutoDetectLanguage: state => state.preferences.spellcheckerAutoDetectLanguage,
-      spellcheckerLanguage: state => state.preferences.spellcheckerLanguage,
-      isHunspellSelected: state => {
-        return !isOsSpellcheckerSupported() || state.preferences.spellcheckerIsHunspell
-      }
+      spellcheckerLanguage: state => state.preferences.spellcheckerLanguage
     })
   },
-  watch: {
-    spellcheckerIsHunspell: function (value, oldValue) {
-      if (this.isOsSpellcheckerSupported && value !== oldValue) {
-        this.ensureDictLanguage(value)
-        this.refreshDictionaryList()
-      }
-    }
-  },
-  created () {
-    this.$nextTick(() => {
-      this.refreshDictionaryList()
-    })
-  },
-  beforeDestroy () {
-    if (!isLinux && this.spellchecker) {
-      this.spellchecker.provider.unsubscribe()
+  mounted () {
+    if (!isOsx) {
+      this.getAvailableDictionaries()
+        .then(dicts => {
+          this.availableDictionaries = dicts
+        })
+
+      ipcRenderer.invoke('mt::spellchecker-get-custom-dictionary-words')
+        .then(words => {
+          this.wordsInCustomDictionary = words.map(word => { return { word } })
+        })
     }
   },
   methods: {
-    getAvailableDictionaries () {
-      let dictionaries = []
-      if (this.isHunspellSelected) {
-        // Search hunspell dictionaries on disk.
-        dictionaries = getAvailableHunspellDictionaries()
-      } else {
-        // We only receive the dictionaries from OS spell checker via the instance.
-        if (!this.spellchecker) {
-          // Create a new spell checker provider without attach it.
-          this.spellchecker = new SpellChecker()
-        }
-
-        // Receive available dictionaries from OS.
-        dictionaries = this.spellchecker.getAvailableDictionaries()
-      }
-
-      return dictionaries.map(item => {
+    async getAvailableDictionaries () {
+      const dictionaries = await SpellChecker.getAvailableDictionaries()
+      return dictionaries.map(selectedItem => {
         return {
-          value: item,
-          label: getLanguageName(item)
+          value: selectedItem,
+          label: getLanguageName(selectedItem)
         }
       })
     },
-    refreshDictionaryList () {
-      this.availableDictionaries = this.getAvailableDictionaries()
-    },
-    ensureDictLanguage (isHunspell) {
-      const { isOsSpellcheckerSupported, spellcheckerLanguage } = this
-      if (isHunspell || !isOsSpellcheckerSupported) {
-        // Validate language for Hunspell.
-        const index = HUNSPELL_DICTIONARY_LANGUAGE_MAP.findIndex(d => d.value === spellcheckerLanguage)
-        if (index === -1) {
-          // Use fallback because language is not supported by Hunspell.
-          this.onSelectChange('spellcheckerLanguage', 'en-US')
-        }
-      } else {
-        // Validate language for OS spellchecker. We only receive the dictionaries from
-        // OS spell checker via the instance.
-        if (!this.spellchecker) {
-          // Create a new spell checker provider without attach it.
-          this.spellchecker = new SpellChecker()
-        }
-
-        const dicts = this.spellchecker.getAvailableDictionaries()
-        const index = dicts.findIndex(d => d === spellcheckerLanguage)
-        if (index === -1 && dicts.length >= 1) {
-          // Language is not supported, prefer OS language.
-          let lang = process.env.LANG
-          lang = lang ? lang.split('.')[0] : null
-          if (lang) {
-            lang = lang.replace(/_/g, '-')
-            if (dicts.findIndex(d => d === lang) === -1) {
-              lang = null
-            }
-          }
-          this.onSelectChange('spellcheckerLanguage', lang || dicts[0])
-        }
+    async ensureDictLanguage (lang) {
+      if (!this.spellchecker) {
+        this.spellchecker = new SpellChecker(true, 'en-US')
       }
+      await this.spellchecker.switchLanguage(lang)
     },
 
-    handleSpellcheckerEnabled (value) {
-      if (value) {
-        const { spellcheckerIsHunspell } = this
-        this.ensureDictLanguage(spellcheckerIsHunspell)
-      }
-      this.onSelectChange('spellcheckerEnabled', value)
+    handleSpellcheckerLanguage (languageCode) {
+      this.ensureDictLanguage(languageCode)
+        .then(() => {
+          this.onSelectChange('spellcheckerLanguage', languageCode)
+        })
+        .catch(error => {
+          log.error(error)
+          notice.notify({
+            title: 'Failed to switch language',
+            type: 'error',
+            message: error.message
+          })
+        })
+    },
+    handleSpellcheckerEnabled (isEnabled) {
+      this.onSelectChange('spellcheckerEnabled', isEnabled)
     },
     onSelectChange (type, value) {
       this.$store.dispatch('SET_SINGLE_PREFERENCE', { type, value })
     },
-
-    // --- Hunspell only ------------------------------------------------------
-
-    addNewDict () {
-      const { selectedDictionaryToAdd } = this
-      if (!this.isHunspellDictionaryAvailable(selectedDictionaryToAdd)) {
-        this.startDownloadHunspellDictionary(selectedDictionaryToAdd)
+    handleDeleteClick (selectedItem) {
+      if (selectedItem && typeof selectedItem.word === 'string') {
+        ipcRenderer.invoke('mt::spellchecker-remove-word', selectedItem.word)
+          .then(success => {
+            if (success) {
+              this.wordsInCustomDictionary = this.wordsInCustomDictionary.filter(item => item.word !== selectedItem.word)
+            } else {
+              notice.notify({
+                title: 'Failed to remove custom word',
+                type: 'error',
+                message: 'An unexpected error occurred while saving.'
+              })
+            }
+          })
+          .catch(error => log.error(error))
       }
-    },
-    handleUpdateClick (index, row) {
-      this.startDownloadHunspellDictionary(row.value)
-    },
-    handleDeleteClick (index, row) {
-      const { spellcheckerLanguage } = this
-      const { value: lang } = row
-
-      // Don't allow to delete our fallback language.
-      if (lang === 'en-US') {
-        return
-      }
-
-      // Fallback before deleting selected language.
-      if (spellcheckerLanguage === lang) {
-        this.onSelectChange('spellcheckerLanguage', 'en-US')
-      }
-
-      deleteHunspellDictionary(lang)
-        .then(() => {
-          this.refreshDictionaryList()
-        }).catch(error => {
-          this.errorMessage = `Error deleting dictionary: ${error.message}`
-        })
-    },
-
-    startDownloadHunspellDictionary (languageCode) {
-      this.errorMessage = ''
-      if (this.hunspellDictionaryDownloadCache[languageCode]) {
-        return
-      } else if (!navigator.onLine) {
-        delete this.hunspellDictionaryDownloadCache[languageCode]
-        this.errorMessage = 'No Internet connection available.'
-        return
-      }
-
-      this.hunspellDictionaryDownloadCache[languageCode] = 1
-      downloadHunspellDictionary(languageCode)
-        .then(() => {
-          delete this.hunspellDictionaryDownloadCache[languageCode]
-          this.refreshDictionaryList()
-        }).catch(error => {
-          delete this.hunspellDictionaryDownloadCache[languageCode]
-          this.errorMessage = `Error while downloading: ${error.message}`
-        })
-    },
-    isHunspellDictionaryAvailable (languageCode) {
-      const { availableDictionaries } = this
-      return availableDictionaries.findIndex(d => d.value === languageCode) !== -1
     }
   }
 }
@@ -306,6 +177,7 @@ export default {
     & h6.title {
       font-weight: 400;
       font-size: 1.1em;
+      margin-bottom: 0;
     }
   }
   .el-table, .el-table__expanded-cell {
@@ -324,28 +196,25 @@ export default {
     background: none;
     border: none;
   }
-  .dictionary-group {
-    display: flex;
-    & button.el-button {
-      height: 30px;
-      width: 30px;
-      padding: 0;
-      margin-left: 6px;
-    }
-
-  }
 </style>
 <style>
   .pref-spellchecker .el-table table {
     margin: 0;
+    border: none;
   }
   .pref-spellchecker .el-table th,
   .pref-spellchecker .el-table tr {
     background: var(--editorBgColor);
   }
-  .pref-spellchecker .el-table td,
-  .pref-spellchecker .el-table th.is-leaf {
-    border: 1px solid var(--tableBorderColor);
+  .pref-spellchecker .el-table th.el-table__cell.is-leaf,
+  .pref-spellchecker .el-table th,
+  .pref-spellchecker .el-table td {
+    border: none;
+  }
+  .pref-spellchecker .el-table th.el-table__cell.is-leaf:last-child,
+  .pref-spellchecker .el-table th:last-child,
+  .pref-spellchecker .el-table td:last-child {
+    border-right: 1px solid var(--tableBorderColor);
   }
   .pref-spellchecker .el-table--border::after,
   .pref-spellchecker .el-table--group::after,
@@ -360,8 +229,12 @@ export default {
   .pref-spellchecker .el-table__body tr.hover-row>td {
     background: var(--selectionColor);
   }
+  .pref-spellchecker .el-table .el-table__cell {
+    padding: 2px 0;
+    margin: 4px 6px;
+  }
 
- .pref-spellchecker li.el-select-dropdown__item {
+  .pref-spellchecker li.el-select-dropdown__item {
     color: var(--editorColor);
     height: 30px;
   }
