@@ -1,6 +1,6 @@
 import { clipboard, ipcRenderer, shell, webFrame } from 'electron'
 import path from 'path'
-import equal from 'deep-equal'
+import equal from 'fast-deep-equal'
 import { isSamePathSync } from 'common/filesystem/paths'
 import bus from '../bus'
 import { hasKeys, getUniqueId } from '../util'
@@ -646,7 +646,7 @@ const actions = {
         showSideBar: !!sideBarVisibility,
         showTabBar: !!tabBarVisibility
       })
-      dispatch('SET_LAYOUT_MENU_ITEM')
+      dispatch('DISPATCH_LAYOUT_MENU_ITEMS')
 
       commit('SET_MODE', {
         type: 'sourceCode',
@@ -701,35 +701,8 @@ const actions = {
   },
 
   LISTEN_FOR_SWITCH_TABS ({ commit, state, dispatch }) {
-    ipcRenderer.on('mt::switch-first-tab', e => {
-      dispatch('SWITCH_TABS', 1)
-    })
-    ipcRenderer.on('mt::switch-second-tab', e => {
-      dispatch('SWITCH_TABS', 2)
-    })
-    ipcRenderer.on('mt::switch-third-tab', e => {
-      dispatch('SWITCH_TABS', 3)
-    })
-    ipcRenderer.on('mt::switch-fourth-tab', e => {
-      dispatch('SWITCH_TABS', 4)
-    })
-    ipcRenderer.on('mt::switch-fifth-tab', e => {
-      dispatch('SWITCH_TABS', 5)
-    })
-    ipcRenderer.on('mt::switch-sixth-tab', e => {
-      dispatch('SWITCH_TABS', 6)
-    })
-    ipcRenderer.on('mt::switch-seventh-tab', e => {
-      dispatch('SWITCH_TABS', 7)
-    })
-    ipcRenderer.on('mt::switch-eighth-tab', e => {
-      dispatch('SWITCH_TABS', 8)
-    })
-    ipcRenderer.on('mt::switch-ninth-tab', e => {
-      dispatch('SWITCH_TABS', 9)
-    })
-    ipcRenderer.on('mt::switch-tenth-tab', e => {
-      dispatch('SWITCH_TABS', 10)
+    ipcRenderer.on('mt::switch-tab-by-index', (event, index) => {
+      dispatch('SWITCH_TAB_BY_INDEX', index)
     })
   },
 
@@ -796,29 +769,30 @@ const actions = {
       console.error(`CYCLE_TABS: Cannot find next tab (index="${nextTabIndex}").`)
       return
     }
+
     commit('SET_CURRENT_FILE', nextTab)
     dispatch('UPDATE_LINE_ENDING_MENU')
   },
 
-  // switch tabs with Alt+#num.
-  SWITCH_TABS ({ commit, dispatch, state }, num) {
+  SWITCH_TAB_BY_INDEX ({ commit, dispatch, state }, nextTabIndex) {
     const { tabs, currentFile } = state
-    if (tabs.length <= 1 || num > tabs.length) {
+    if (nextTabIndex < 0 || nextTabIndex >= tabs.length) {
+      console.warn('Invalid tab index:', nextTabIndex)
       return
     }
 
     const currentIndex = tabs.findIndex(t => t.id === currentFile.id)
     if (currentIndex === -1) {
-      console.error('CYCLE_TABS: Cannot find current tab index.')
+      console.error('Cannot find current tab index.')
       return
     }
 
-    const nextTabIndex = num - 1
     const nextTab = tabs[nextTabIndex]
     if (!nextTab || !nextTab.id) {
-      console.error(`CYCLE_TABS: Cannot find next tab (index="${nextTabIndex}").`)
+      console.error(`Cannot find tab by index="${nextTabIndex}".`)
       return
     }
+
     commit('SET_CURRENT_FILE', nextTab)
     dispatch('UPDATE_LINE_ENDING_MENU')
   },
@@ -916,7 +890,7 @@ const actions = {
     const { tabs } = state
     if (always || tabs.length === 1) {
       commit('SET_LAYOUT', { showTabBar: true })
-      dispatch('SET_LAYOUT_MENU_ITEM')
+      dispatch('DISPATCH_LAYOUT_MENU_ITEMS')
     }
   },
 
@@ -1231,8 +1205,32 @@ const actions = {
   },
 
   LISTEN_FOR_RELOAD_IMAGES () {
-    ipcRenderer.on('mt::invalidate-image-cache', (e) => {
+    ipcRenderer.on('mt::invalidate-image-cache', () => {
       bus.$emit('invalidate-image-cache')
+    })
+  },
+
+  LISTEN_FOR_CONTEXT_MENU () {
+    // General context menu
+    ipcRenderer.on('mt::cm-copy-as-markdown', () => {
+      bus.$emit('copyAsMarkdown', 'copyAsMarkdown')
+    })
+    ipcRenderer.on('mt::cm-copy-as-html', () => {
+      bus.$emit('copyAsHtml', 'copyAsHtml')
+    })
+    ipcRenderer.on('mt::cm-paste-as-plain-text', () => {
+      bus.$emit('pasteAsPlainText', 'pasteAsPlainText')
+    })
+    ipcRenderer.on('mt::cm-insert-paragraph', (e, location) => {
+      bus.$emit('insertParagraph', location)
+    })
+
+    // Spelling
+    ipcRenderer.on('mt::spelling-replace-misspelling', (e, info) => {
+      bus.$emit('replace-misspelling', info)
+    })
+    ipcRenderer.on('mt::spelling-show-switch-language', () => {
+      bus.$emit('open-command-spellchecker-switch-language')
     })
   }
 }
