@@ -33,7 +33,7 @@ const mutations = {
 }
 
 const actions = {
-  LISTEN_FOR_LAYOUT ({ state, commit }) {
+  LISTEN_FOR_LAYOUT ({ state, commit, dispatch }) {
     ipcRenderer.on('mt::set-view-layout', (e, layout) => {
       if (layout.rightColumn) {
         commit('SET_LAYOUT', {
@@ -44,26 +44,27 @@ const actions = {
       } else {
         commit('SET_LAYOUT', layout)
       }
+      dispatch('DISPATCH_LAYOUT_MENU_ITEMS')
     })
 
-    bus.$on('view:toggle-view-layout-entry', entryName => {
+    ipcRenderer.on('mt::toggle-view-layout-entry', (event, entryName) => {
       commit('TOGGLE_LAYOUT_ENTRY', entryName)
-      const item = {}
-      item[entryName] = state[entryName]
+      dispatch('DISPATCH_LAYOUT_MENU_ITEMS')
+    })
+
+    bus.$on('view:toggle-layout-entry', entryName => {
+      commit('TOGGLE_LAYOUT_ENTRY', entryName)
       const { windowId } = global.marktext.env
-      ipcRenderer.send('mt::view-layout-changed', windowId, item)
+      ipcRenderer.send('mt::view-layout-changed', windowId, { [entryName]: state[entryName] })
     })
   },
-  LISTEN_FOR_REQUEST_LAYOUT ({ dispatch }) {
-    ipcRenderer.on('mt::request-for-view-layout', () => {
-      dispatch('SET_LAYOUT_MENU_ITEM')
-    })
-  },
-  SET_LAYOUT_MENU_ITEM ({ state }) {
+
+  DISPATCH_LAYOUT_MENU_ITEMS ({ state }) {
     const { windowId } = global.marktext.env
     const { showTabBar, showSideBar } = state
     ipcRenderer.send('mt::view-layout-changed', windowId, { showTabBar, showSideBar })
   },
+
   CHANGE_SIDE_BAR_WIDTH ({ commit }, width) {
     commit('SET_SIDE_BAR_WIDTH', width)
   }
